+++ chips/nrf5x-unsafe/README.md
@@ -1,0 +1,3 @@
+# Nordic Semiconductor nRF5X Series SoC - Unsafe
+
+This contains the `unsafe` code for the nrf5x chip.
